moved search feature into core
