+++ shared/src/commonMain/kotlin/de/awolf/trip/kmp/core/data/remote/HttpRoutes.kt
@@ -1,0 +1,6 @@
+package de.awolf.trip.kmp.core.data.remote
+
+object HttpRoutes {
+    private const val BASE_URL = "https://webapi.vvo-online.de"
+    const val STOP_SEARCH = "$BASE_URL/tr/pointfinder"
+}
