+++ shared/src/commonMain/kotlin/de/awolf/trip/kmp/core/domain/models/StopListSource.kt
@@ -1,0 +1,7 @@
+package de.awolf.trip.kmp.core.domain.models
+
+enum class StopListSource {
+    NONE,
+    FAVORITES,
+    SEARCH,
+}
