+++ shared/src/commonMain/kotlin/de/awolf/trip/kmp/core/domain/use_cases/SearchUseCases.kt
@@ -1,0 +1,8 @@
+package de.awolf.trip.kmp.core.domain.use_cases
+
+data class SearchUseCases(
+    val findStopByQuery: FindStopByQueryUseCase,
+    val getFavoriteStops: GetFavoriteStopsUseCase,
+    val reorderFavoriteStops: ReorderFavoriteStopsUseCase,
+    val toggleFavoriteStop: ToggleFavoriteStopUseCase,
+)
