--- shared/src/commonMain/kotlin/de/awolf/trip/kmp/departures/presentation/DepartureDetailLevel.kt
@@ -1,8 +1,0 @@
-package de.awolf.trip.kmp.departures.presentation
-
-enum class DepartureDetailLevel {
-    NONE,
-    PLATFORM,
-    STOP_SCHEDULE,
-    WHOLE_STOP_SCHEDULE,
-}
