+++ shared/src/commonMain/kotlin/de/awolf/trip/kmp/departures/presentation/departures_screen/DepartureDetailLevel.kt
@@ -1,0 +1,8 @@
+package de.awolf.trip.kmp.departures.presentation.departures_screen
+
+enum class DepartureDetailLevel {
+    NONE,
+    PLATFORM,
+    STOP_SCHEDULE,
+    WHOLE_STOP_SCHEDULE,
+}
