+++ shared/src/commonMain/kotlin/de/awolf/trip/kmp/departures/presentation/departures_screen/DeparturesScreenSideEffect.kt
@@ -1,0 +1,7 @@
+package de.awolf.trip.kmp.departures.presentation.departures_screen
+
+import de.awolf.trip.kmp.core.util.error.NetworkError
+
+sealed interface DeparturesScreenSideEffect {
+    data class ShowNetworkError(val error: NetworkError) : DeparturesScreenSideEffect
+}
