--- shared/src/commonMain/kotlin/de/awolf/trip/kmp/search/data/local/createStopDatabase.kt
@@ -1,10 +1,0 @@
-package de.awolf.trip.kmp.search.data.local
-
-import app.cash.sqldelight.db.SqlDriver
-import de.awolf.trip.kmp.shared.database.StopDatabase
-
-fun createStopDatabase(driver: SqlDriver): StopDatabase {
-    return StopDatabase(
-        driver = driver
-    )
-}
