--- shared/src/commonMain/kotlin/de/awolf/trip/kmp/search/data/remote/HttpRoutes.kt
@@ -1,6 +1,0 @@
-package de.awolf.trip.kmp.search.data.remote
-
-object HttpRoutes {
-    private const val BASE_URL = "https://webapi.vvo-online.de"
-    const val STOP_SEARCH = "$BASE_URL/tr/pointfinder"
-}
