--- shared/src/commonMain/kotlin/de/awolf/trip/kmp/search/domain/models/StopListSource.kt
@@ -1,7 +1,0 @@
-package de.awolf.trip.kmp.search.domain.models
-
-enum class StopListSource {
-    NONE,
-    FAVORITES,
-    SEARCH,
-}
