--- shared/src/commonMain/kotlin/de/awolf/trip/kmp/search/domain/repository/SearchRemoteRepository.kt
@@ -1,15 +1,0 @@
-package de.awolf.trip.kmp.search.domain.repository
-
-import de.awolf.trip.kmp.core.util.Result
-import de.awolf.trip.kmp.core.util.error.NetworkError
-import de.awolf.trip.kmp.search.domain.models.StopFinderInfo
-
-interface SearchRemoteRepository {
-    suspend fun getStopByName(
-        query: String,
-        limit: Int,
-        stopsOnly: Boolean,
-        regionalOnly: Boolean,
-        stopShortcuts: Boolean
-    ): Result<StopFinderInfo, NetworkError>
-}
