+++ shared/src/iosMain/kotlin/de/awolf/trip/kmp/di/Modules.ios.kt
@@ -1,7 +1,7 @@
 package de.awolf.trip.kmp.di
 
 import app.cash.sqldelight.driver.native.NativeSqliteDriver
-import de.awolf.trip.kmp.search.data.local.createStopDatabase
+import de.awolf.trip.kmp.core.data.local.createStopDatabase
 import de.awolf.trip.kmp.core.data.remote.createHttpClient
 import de.awolf.trip.kmp.shared.database.StopDatabase
 import io.ktor.client.HttpClient
